+++ src/java/au/com/dius/emr/CommonCrawlReducer.java
@@ -1,16 +1,17 @@
 package au.com.dius.emr;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
-public class CommonCrawlReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
+public class CommonCrawlReducer extends Reducer<Text, IntWritable, Text, LongWritable> {
 
-  public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
-    int sum = 0;
-    for (LongWritable val : values) {
+  public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+    long sum = 0;
+    for (IntWritable val : values) {
       sum += val.get();
     }
     context.write(key, new LongWritable(sum));
